requirements.txt
# Required Python Libraries

# For creating and managing directories and files
os

# For parsing and handling JSON data
json

# For handling command-line arguments and user prompts
argparse

# For AI-Optimized Organization
tensorflow>=2.4.1
numpy>=1.19.5
pandas>=1.2.1

# For Integration with Version Control
GitPython>=3.1.14

# For generating documentation
pydoc-markdown>=3.11.0
